update 上游更新 20260427

+++ app/src/main/java/io/legado/app/model/localBook/PdfFile.kt
@@ -4,6 +4,7 @@ import android.graphics.Bitmap
 import android.graphics.Color
 import android.graphics.pdf.PdfRenderer
 import android.os.ParcelFileDescriptor
+import androidx.core.graphics.createBitmap
 import io.legado.app.constant.AppLog
 import io.legado.app.data.entities.Book
 import io.legado.app.data.entities.BookChapter
@@ -74,6 +75,10 @@ class PdfFile(var book: Book) {
             return field
         }
 
+    init {
+        upBookCover(true)
+    }
+
     /**
      * 读取PDF文件
      *
@@ -117,16 +122,14 @@ class PdfFile(var book: Book) {
         if (index >= renderer.pageCount) {
             return null
         }
-        return renderer.openPage(index)?.use { page ->
-            Bitmap.createBitmap(
+        return renderer.openPage(index).use { page ->
+            createBitmap(
                 SystemUtils.screenWidthPx,
-                (SystemUtils.screenWidthPx.toDouble() * page.height / page.width).toInt(),
-                Bitmap.Config.ARGB_8888
-            )
-                .apply {
-                    this.eraseColor(Color.WHITE)
-                    page.render(this, null, null, PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY)
-                }
+                (SystemUtils.screenWidthPx.toDouble() * page.height / page.width).toInt()
+            ).apply {
+                this.eraseColor(Color.WHITE)
+                page.render(this, null, null, PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY)
+            }
         }
 
     }
@@ -164,7 +167,7 @@ class PdfFile(var book: Book) {
                 null
             }
 
-        } catch (e: Exception) {
+        } catch (_: Exception) {
             return null
         }
     }
@@ -188,11 +191,14 @@ class PdfFile(var book: Book) {
         return chapterList
     }
 
-    private fun upBookCover() {
+    private fun upBookCover(fastCheck: Boolean = false) {
         try {
             pdfRenderer?.let { renderer ->
                 if (book.coverUrl.isNullOrEmpty()) {
                     book.coverUrl = LocalBook.getCoverPath(book)
+                }
+                if (fastCheck && File(book.coverUrl!!).exists()) {
+                    return
                 }
                 FileOutputStream(FileUtils.createFileIfNotExist(book.coverUrl!!)).use { out ->
                     openPdfPage(renderer, 0)?.compress(Bitmap.CompressFormat.JPEG, 90, out)
